+++ packages/screenshoter/src/scroller.js
@@ -80,7 +80,6 @@ function makeScroller({logger, element, scrollingMode = 'mixed'}) {
 
   async function scrollTo(offset, element = defaultElement) {
     try {
-      // offset = {x: Math.max(offset.x, 0), y: Math.max(offset.y, 0)}
       const scrollOffset = await element.scrollTo(offset)
       return scrollOffset
     } catch (err) {
@@ -92,7 +91,6 @@ function makeScroller({logger, element, scrollingMode = 'mixed'}) {
 
   async function translateTo(offset, element = defaultElement) {
     try {
-      // offset = {x: Math.max(offset.x, 0), y: Math.max(offset.y, 0)}
       await element.scrollTo({x: 0, y: 0})
       const translateOffset = await element.translateTo(offset)
       return translateOffset
@@ -105,8 +103,15 @@ function makeScroller({logger, element, scrollingMode = 'mixed'}) {
 
   async function shiftTo(offset, element = defaultElement) {
     try {
-      // offset = {x: Math.max(offset.x, 0), y: Math.max(offset.y, 0)}
       const scrollOffset = await element.scrollTo(offset)
+      if (utils.geometry.equals(scrollOffset, offset)) return scrollOffset
+
+      // there is a "bug" in iOS that will not move a root element if it already scrolled, so it should be translated all the way
+      if (element.driver.isIOS && (await element.isRoot())) {
+        const translateOffset = await element.translateTo(offset)
+        return translateOffset
+      }
+
       const remainingOffset = utils.geometry.offsetNegative(offset, scrollOffset)
       const translateOffset = await element.translateTo(remainingOffset)