+++ packages/snippets/src/browser/getViewportScale.js
@@ -1,0 +1,5 @@
+function getViewportScale() {
+  return window.visualViewport && window.visualViewport.scale
+}
+
+module.exports = getViewportScale