+++ packages/snippets/src/browser/getViewportSize.js
@@ -16,11 +16,6 @@ function getViewportSize() {
     width = document.body.clientWidth
   }
 
-  if (window.visualViewport) {
-    width = Math.round(width * window.visualViewport.scale)
-    height = Math.round(height * window.visualViewport.scale)
-  }
-
   return {width, height}
 }